Reactoring WindowsActivityTracker continued
